ai-service/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 计算机视觉和机器学习
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2

# 深度学习框架
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0

# 数据处理
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# 图像处理
imageio==2.31.6
scikit-image==0.22.0

# 工具库
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 日志和监控
loguru==0.7.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 其他依赖
requests==2.31.0
websockets==12.0
redis==5.0.1
celery==5.3.4
